arp: share ARP-reply BPF filter across linux and macOS
